feat(mcp): add SSE transport unified behind McpConnection trait
- Extract the transport-agnostic McpServerManager, JSON-RPC message
  types, and error type out of mcp_stdio.rs into mcp_server_manager.rs
  (structural refactor, no logic change); route the SSE transport
  through the manager's spawn path alongside stdio.
- Add the McpConnection trait abstracting the JSON-RPC connection
  surface so the manager drives stdio and SSE through a single
  Box<dyn McpConnection>.
- Implement the legacy HTTP+SSE transport (mcp_sse.rs) on reqwest: open
  a long-lived GET stream, resolve the `endpoint` event, POST JSON-RPC
  requests, and read responses off the stream; support headersHelper
  for dynamic per-request headers.
- Add reqwest (rustls-tls + stream) as a runtime dependency for the
  SSE HTTP client; update imports/re-exports in lib.rs, mcp_server.rs,
  and mcp_tool_bridge.rs.
